Move encode to files, remove if/else
