+++ genai-perf/genai_perf/inputs/retrievers/input_retriever_factory.py
@@ -26,16 +26,12 @@
 
 from typing import Dict, List
 
-from genai_perf import utils
 from genai_perf.exceptions import GenAIPerfException
 from genai_perf.inputs.input_constants import OutputFormat, PromptSource
 from genai_perf.inputs.inputs_config import InputsConfig
 from genai_perf.inputs.retrievers.file_input_retriever import FileInputRetriever
 from genai_perf.inputs.retrievers.generic_dataset import GenericDataset
 from genai_perf.inputs.retrievers.synthetic_data_retriever import SyntheticDataRetriever
-from genai_perf.inputs.retrievers.synthetic_image_generator import ImageFormat
-from PIL import Image
-from requests import Response
 
 
 class InputRetrieverFactory:
@@ -52,29 +48,16 @@ class InputRetrieverFactory:
             The generic dataset JSON
         """
 
-        input_data: GenericDataset = None
-        if self.config.output_format in [
-            OutputFormat.RANKINGS,
-            OutputFormat.IMAGE_RETRIEVAL,
-        ]:
+        # TODO: remove once the factory fully integrates retrievers
+        if self.config.input_type == PromptSource.SYNTHETIC:
+            synthetic_retriever = SyntheticDataRetriever(self.config)
+            input_data = synthetic_retriever.retrieve_data()
+        elif self.config.input_type == PromptSource.FILE:
             # TODO: remove once the factory fully integrates retrievers
             file_retriever = FileInputRetriever(self.config)
             input_data = file_retriever.retrieve_data()
-
-            if self.config.output_format == OutputFormat.IMAGE_RETRIEVAL:
-                input_data = self._encode_images_in_input_dataset(input_data)
-
         else:
-            if self.config.input_type == PromptSource.SYNTHETIC:
-                synthetic_retriever = SyntheticDataRetriever(self.config)
-                input_data = synthetic_retriever.retrieve_data()
-            elif self.config.input_type == PromptSource.FILE:
-                # TODO: remove once the factory fully integrates retrievers
-                file_retriever = FileInputRetriever(self.config)
-                input_data = file_retriever.retrieve_data()
-                self._encode_images_in_input_dataset(input_data)
-            else:
-                raise GenAIPerfException("Input source is not recognized.")
+            raise GenAIPerfException("Input source is not recognized.")
 
         return input_data
 
@@ -84,33 +67,6 @@ class InputRetrieverFactory:
         generic_dataset_json = self._convert_dataset_to_generic_input_json(dataset)
 
         return generic_dataset_json
-
-    def _encode_images_in_input_dataset(self, input_file_dataset: GenericDataset):
-        for file_data in input_file_dataset.files_data.values():
-            for row in file_data.rows:
-                for i, image in enumerate(row.images):
-                    if image is not None:
-                        payload = self._encode_image(image)
-                        row.images[i] = payload
-
-    def _encode_image(self, filename: str) -> str:
-        img = Image.open(filename)
-        if img is None:
-            raise GenAIPerfException(f"Failed to open image '{filename}'.")
-        if img.format is None:
-            raise GenAIPerfException(
-                f"Failed to determine image format of '{filename}'."
-            )
-
-        if img.format.lower() not in utils.get_enum_names(ImageFormat):
-            raise GenAIPerfException(
-                f"Unsupported image format '{img.format}' of "
-                f"the image '{filename}'."
-            )
-
-        img_base64 = utils.encode_image(img, img.format)
-        payload = f"data:image/{img.format.lower()};base64,{img_base64}"
-        return payload
 
     def _check_for_error_in_json_of_dataset(self, dataset_json: Dict) -> None:
         if "error" in dataset_json:
