+++ genai-perf/rankings_jsonl/queries.jsonl
@@ -1,0 +1,4 @@
+{"text": "What was the first car ever driven?"}
+{"text": "Who served as the 5th President of the United States of America?"}
+{"text": "Is the Sydney Opera House located in Australia?"}
+{"text": "In what state did they film Shrek 2?"}
